requirements.txt
openai>=1.0.0
google-genai>=0.8.0
anthropic>=0.18.0
matplotlib>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
